moved tests to separate file
